+++ Tavernspiel/src/gui/mainToolbox/hudLayout/DefaultHUDLookAndFeel.java
@@ -80,7 +80,7 @@ public class DefaultHUDLookAndFeel implements HUDStrategy{
         g.fill3DRect(Game.WIDTH - 70, Game.HEIGHT - 300, 70, 40, true);
         
         g.setColor(Color.red);
-        g.fill3DRect(70, 5, (int) ((double)Window.main.player.attributes.hp/(double)(Window.main.player.attributes.maxhp) * 200), 10, true);
+        g.fill3DRect(70, 5, (int) (Window.main.player.attributes.hp/(double)(Window.main.player.attributes.maxhp) * 200), 10, true);
         
         int x = 29, y = 77;
         for(Buff b : hero.buffs){
